Update expected figures to use matplotlib 3.5 (#590)
* Change to `pytest` instead of `py.test`

* Silence a warning

* Update test images to use matplotlib 3.5

* Add to release notes

* Set "older" matplotlib version in dependency-tests to current version

+++ pyam/plotting.py
@@ -437,7 +437,7 @@ def stack(
 
     def as_series(index, name):
         _idx = [i[0] for i in index]
-        return pd.Series([0] * len(index), index=_idx, name=name)
+        return pd.Series([0] * len(index), index=_idx, name=name, dtype="int")
 
     # determine all time-indices where a timeseries crosses 0 and add to data
     _rows = pd.concat(
